Add DataExtraction from all accounts of user

+++ db/migrate/20240122064113_create_ad_campaigns.rb
@@ -5,7 +5,7 @@ class CreateAdCampaigns < ActiveRecord::Migration[7.1]
       t.string        :name
       t.string        :objective
       t.datetime      :startdate
-      t.integer       :lifetime_budget
+      t.bigint        :lifetime_budget
       t.string        :budgeting_type
       t.string        :account_id, null: false
       t.timestamps
